django/requirements.txt
certifi==2017.4.17
chardet==3.0.4
construct==2.5.3
Django==1.11.3
django-bootstrap3==8.2.3
django-enumfields==0.9.0
django-haystack==2.6.1
django-jquery==3.1.0
django-modeltranslation==0.12.1
django-rest-framework==0.1.0
django-taggit==0.22.1
django-tagulous==0.12.0
djangorestframework==3.6.3
feedparser==5.2.1
future==0.16.0
haystack==0.42
httplib2==0.10.3
idna==2.5
pathlib==1.0.1
pefile==2017.9.3
psycopg2==2.7.3.1
pysolr==3.6.0
python-ptrace==0.9.3
pytz==2017.2
Pyzotero==1.2.13
requests==2.18.1
six==1.11.0
urllib3==1.21.1
